+++ src/chart/rendering/drawinterlacing.cpp
@@ -34,7 +34,9 @@ void DrawInterlacing::drawGeometries(Gen::AxisId axisIndex) const
 	    || guides.interlacings == false)
 		return;
 
-	auto orientation = !+axisIndex;
+	auto otherWeights = getInterlacingWeights(!axisIndex);
+	auto &&otherInterlacingColor =
+	    *parent.rootStyle.plot.getAxis(!axisIndex).interlacing.color;
 
 	parent.painter.setPolygonToCircleFactor(0);
 	parent.painter.setPolygonStraightFactor(0);
@@ -48,30 +50,42 @@ void DrawInterlacing::drawGeometries(Gen::AxisId axisIndex) const
 		                       1.0,
 		                       Math::Floating::less)
 		                 - clippedBottom;
-		auto rect = Geom::Rect{
-		    Geom::Point::Coord(orientation, 0.0, clippedBottom),
-		    {Geom::Size::Coord(orientation, 1.0, clippedSize)}};
 
-		auto interlacingColor =
-		    *axisStyle.interlacing.color
-		    * Math::FuzzyBool::And<double>(interval.weight,
-		        interval.isSecond,
-		        guides.interlacings);
+		auto rect = [&, orientation = +axisIndex](const double &from,
+		                const double &to)
+		{
+			return Geom::Rect{
+			    Geom::Point::Coord(orientation, clippedBottom, from),
+			    {Geom::Size::Coord(orientation,
+			        clippedSize,
+			        to - from)}};
+		};
 
-		auto &canvas = parent.canvas;
-		canvas.save();
-		canvas.setLineColor(Gfx::Color::Transparent());
-		canvas.setBrushColor(interlacingColor);
-		if (auto &&eventTarget =
-		        Events::Targets::axisInterlacing(axisIndex);
-		    parent.rootEvents.draw.plot.axis.interlacing->invoke(
-		        Events::OnRectDrawEvent(*eventTarget,
-		            {rect, true}))) {
-			parent.painter.drawPolygon(rect.points());
-			parent.renderedChart.emplace(Rect{rect, true},
-			    std::move(eventTarget));
+		auto weight = Math::FuzzyBool::And<double>(interval.weight,
+		    interval.isSecond,
+		    guides.interlacings);
+		auto interlacingColor = *axisStyle.interlacing.color * weight;
+
+		for (auto first = otherWeights.begin(),
+		          next = std::next(first),
+		          last = otherWeights.end();
+		     next != last;
+		     ++next, ++first) {
+			if (Math::Floating::is_zero(first->second))
+				drawInterlacing(axisIndex,
+				    interlacingColor,
+				    rect(first->first, next->first));
+			else if (axisIndex == Gen::AxisId::y) {
+				drawInterlacing(first->second > weight ? !axisIndex
+				                                       : axisIndex,
+				    getCrossingInterlacingColor(
+				        *axisStyle.interlacing.color,
+				        weight,
+				        otherInterlacingColor,
+				        first->second),
+				    rect(first->first, next->first));
+			}
 		}
-		canvas.restore();
 	}
 }
 
@@ -117,6 +131,26 @@ void DrawInterlacing::drawTexts(Gen::AxisId axisIndex) const
 			    axisIndex,
 			    tickPos);
 	}
+}
+
+void DrawInterlacing::drawInterlacing(Gen::AxisId axisIndex,
+    const Gfx::Color &interlacingColor,
+    const Geom::Rect &rect) const
+{
+	auto &canvas = parent.canvas;
+	canvas.save();
+	canvas.setLineColor(Gfx::Color::Transparent());
+	canvas.setLineWidth(0);
+	canvas.setBrushColor(interlacingColor);
+	if (auto &&eventTarget =
+	        Events::Targets::axisInterlacing(axisIndex);
+	    parent.rootEvents.draw.plot.axis.interlacing->invoke(
+	        Events::OnRectDrawEvent(*eventTarget, {rect, true}))) {
+		parent.painter.drawPolygon(rect.points());
+		parent.renderedChart.emplace(Rect{rect, true},
+		    std::move(eventTarget));
+	}
+	canvas.restore();
 }
 
 void DrawInterlacing::drawDataLabel(
@@ -224,6 +258,63 @@ void DrawInterlacing::drawSticks(double tickLength,
 	}
 
 	canvas.restore();
+}
+
+std::map<double, double> DrawInterlacing::getInterlacingWeights(
+    Gen::AxisId axisIndex) const
+{
+	std::map<double, double> weights{{0.0, 0.0}, {1.0, 0.0}};
+
+	auto &&guides = parent.plot->guides.at(axisIndex);
+	auto &&axisStyle = parent.rootStyle.plot.getAxis(axisIndex);
+	if (axisStyle.interlacing.color->isTransparent()
+	    || guides.interlacings == false)
+		return weights;
+
+	for (auto &&interval : parent.getIntervals(axisIndex)) {
+		if (Math::Floating::is_zero(interval.isSecond)) continue;
+		auto min = std::max(interval.range.getMin(), 0.0);
+		auto max = std::min(interval.range.getMax(), 1.0);
+		auto mprev = std::prev(weights.upper_bound(min));
+		auto mnext = weights.lower_bound(max);
+
+		if (mprev->first < min)
+			mprev = weights.try_emplace(mprev, min, mprev->second);
+		if (mnext->first > max)
+			mnext = weights.try_emplace(mnext,
+			    max,
+			    std::prev(mnext)->second);
+
+		while (mprev != mnext)
+			mprev++->second +=
+			    Math::FuzzyBool::And<double>(interval.weight,
+			        interval.isSecond,
+			        guides.interlacings);
+	}
+	return weights;
+}
+
+Gfx::Color DrawInterlacing::getCrossingInterlacingColor(
+    const Gfx::Color &mainColor,
+    double mainWeight,
+    const Gfx::Color &otherColor,
+    double otherWeight)
+{
+	auto color = mainColor * mainWeight + otherColor * otherWeight;
+
+	color.alpha = 1
+	            - (1 - mainColor.alpha * mainWeight)
+	                  * (1 - otherColor.alpha * otherWeight);
+
+	if (mainWeight + otherWeight > 1.0)
+		color = Math::Niebloid::interpolate(color,
+		    color
+		        * std::max({std::abs(mainColor.red - otherColor.red),
+		            std::abs(mainColor.green - otherColor.green),
+		            std::abs(mainColor.blue - otherColor.blue)}),
+		    mainWeight + otherWeight - 1.0);
+
+	return color;
 }
 
 }